+++ docs/.vitepress/components/form/align.vue
@@ -38,3 +38,9 @@
         </FTabPane>
     </FTabs>
 </template>
+
+<style>
+.tab-content {
+    margin-top: 24px;
+}
+</style>